pyproject.toml
[project]
name = "proj"
version = "0.1.0"
description = ""
authors = [
    {name = "Bruce Ruff"}
]
license = "MIT License"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "numpy (>=2.4.4,<3.0.0)",
    "pandas (>=3.0.2,<4.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "plotly (>=6.6.0,<7.0.0)",
    "scipy (>=1.17.1,<2.0.0)",
    "neuprint-python (>=0.6.1,<0.7.0)",
    "scikit-learn (>=1.8.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
